+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-addopts = --doctest-modules --mypy
+addopts = --doctest-modules --mypy --ignore-glob=tests/bindgen/*/app.py
 norecursedirs = 
-    tests/bindgen/*
+    tests/bindgen/generated/*
